+++ libs/shared/src/index.ts
@@ -1,11 +1,16 @@
 export * from './lib/auth';
 export * from './lib/components';
+export * from './lib/datasets';
 export { ExternalContentComponent, externalContentRoutes } from './lib/features/external-content';
 export {
   OperationsHistoryChildListComponent,
   OperationsHistoryModalComponent,
 } from './lib/features/operations-history';
-export { ServicesHistoryChildListComponent, ServicesHistoryViewComponent } from './lib/features/services-history';
+export {
+  ServicesHistoryChildListComponent,
+  ServicesHistoryFilterComponent,
+  ServicesHistoryViewComponent,
+} from './lib/features/services-history';
 export * from './lib/i18n';
 export * from './lib/layouts';
 export * from './lib/models';